terminado modulo productos

+++ src/app/modules/shared/components/confirm/confirm.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, inject } from '@angular/core';
 import { CategoryService } from '../../services/category.service';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { ProductService } from '../../service/product.service';
 
 @Component({
   selector: 'app-confirm',
@@ -12,7 +13,7 @@ export class ConfirmComponent implements OnInit{
   private categoryService = inject(CategoryService);
   private dialogRef = inject(MatDialogRef);
   public data = inject(MAT_DIALOG_DATA);
-  
+  private productService = inject(ProductService);
   ngOnInit(): void {
     
   }
@@ -25,19 +26,35 @@ export class ConfirmComponent implements OnInit{
    */
   delete(){
     if(this.data!=null){
-      this.categoryService.deleteCategory(this.data.id)
-      .subscribe({
-        next: (data:any) =>{
-        this.dialogRef.close(1);
-        }, 
-        error: (error:any)=>{
+      
+      if(this.data.module == "category"){
+        this.categoryService.deleteCategory(this.data.id)
+          .subscribe({
+            next: (data:any) =>{
+            this.dialogRef.close(1);
+            }, 
+            error: (error:any)=>{
+            this.dialogRef.close(2);
+            }
+          })
+      } else {
         this.dialogRef.close(2);
-        }
       }
-      )}
-     else {
-      this.dialogRef.close(2);
-    }
+
+      if(this.data.module == "product"){
+        this.productService.deleteProduct(this.data.id)
+          .subscribe({
+            next: (data:any) =>{
+            this.dialogRef.close(1);
+            }, 
+            error: (error:any)=>{
+            this.dialogRef.close(2);
+            }
+          })
+      } else {
+        this.dialogRef.close(2);
+      }
+    }  
   }
 }
